pyproject.toml
[project]
name = "abucitiserv"
version = "0.1.0"
description = ""
authors = [
    {name = "Avanish Gupta",email = "avanish.sde@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "minio (>=7.2.20,<8.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.14"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
